moved load balancing classes and base classes into a more organized file and directory structure.
