+++ pox/misc/loadbalancing/__init__.py
@@ -1,0 +1,3 @@
+"""
+The components in this package are load balancing classes fully usable with POX controller.
+"""
