+++ pox/misc/loadbalancing/base/__init__.py
@@ -1,0 +1,4 @@
+"""
+The components in this package contain base classes that contain common methods used by
+loadbalancing subclasses, such as lb_dest_hashing.
+"""
